Move top-level render functions to "shell" modules
