+++ packages/site-client/src/entrypoints/element.ts
@@ -4,9 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import '../pages/element/wco-element-page.js';
-import type {ElementData} from '../pages/element/wco-element-page.js';
-import {html} from 'lit';
-
-export const renderElementPage = (elementData: ElementData) =>
-  html`<wco-element-page .elementData=${elementData}></wco-element-page>`;
+export * from '../pages/element/shell.js';
